+++ lib/data_go_kr.dart
@@ -3,4 +3,5 @@
 library;
 
 export 'src/client.dart';
+export 'src/kma_grid.dart';
 export 'src/response.dart';